// src/Client/src/Gen/models/PartResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BikeService.Server | v1
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PartResponse
 */
export interface PartResponse {
    /**
     * 
     * @type {string}
     * @memberof PartResponse
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof PartResponse
     */
    manufacturer: string;
    /**
     * 
     * @type {string}
     * @memberof PartResponse
     */
    model: string;
    /**
     * 
     * @type {number}
     * @memberof PartResponse
     */
    mileage: number;
    /**
     * 
     * @type {string}
     * @memberof PartResponse
     */
    typeId: string;
    /**
     * 
     * @type {string}
     * @memberof PartResponse
     */
    bikeId: string;
}

/**
 * Check if a given object implements the PartResponse interface.
 */
export function instanceOfPartResponse(value: object): value is PartResponse {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('manufacturer' in value) || value['manufacturer'] === undefined) return false;
    if (!('model' in value) || value['model'] === undefined) return false;
    if (!('mileage' in value) || value['mileage'] === undefined) return false;
    if (!('typeId' in value) || value['typeId'] === undefined) return false;
    if (!('bikeId' in value) || value['bikeId'] === undefined) return false;
    return true;
}

export function PartResponseFromJSON(json: any): PartResponse {
    return PartResponseFromJSONTyped(json, false);
}

export function PartResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): PartResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'manufacturer': json['manufacturer'],
        'model': json['model'],
        'mileage': json['mileage'],
        'typeId': json['typeId'],
        'bikeId': json['bikeId'],
    };
}

export function PartResponseToJSON(json: any): PartResponse {
    return PartResponseToJSONTyped(json, false);
}

export function PartResponseToJSONTyped(value?: PartResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'manufacturer': value['manufacturer'],
        'model': value['model'],
        'mileage': value['mileage'],
        'typeId': value['typeId'],
        'bikeId': value['bikeId'],
    };
}